rdkit Properties learn how to be const-correct (logically and physically)
This has a large ramification on the code-base.
One of the issues is that in C++11 and forward,
at least on Microsoft C++, const really does
have a strict meaning especially for stl and
threading:

http://herbsutter.com/2013/01/01/video-you-dont-know-const-and-mutable/

Basically one of the definitions of const is:

const &v means either one of the following:

1. v doesn’t change (“bitwise immutable”)
2. changes to v are synchronized against reads
   from other threads

In the original implementation (v, i.e. an atom
that pointed to a property dict) didn’t change,
however it was NOT synchronized against reads
from other threads so the const designation
was invalid.

In the current implementation, neither (1) nor (2)
are true.

In any case neither setProp nor clearProp can
be considered const.

Our usage of get/set Prop was certainly lazy
caching of data, however to users of the
interface, this is not obvious and the
definition of const (i.e. chemically const)
is not a language definition but rather a
domain definition.