Better wrapping of RWMol
To call RWMol& based functions we need to expose the RWMol
bases otherwise boost doesn’t know about them.  To do this
we make a “Fake” class that really can’t be used.

The ReadWriteMol class is then used as the real class.  This
has a side benefit that allows us to fix an annoying boost
issue for overloads:

  Foo(const ROMol&)
  Foo(RWMol&)

When called form boost, boost::python finds a proper conversion
from an romol to a temporary rwmol ala:

  >>> m = Chem.Mol()
  >>> Chem.Foo(m) # calls Foo(RWMol&) via a temporary

Yuck.  Using the boost wrapping of:

  Foo(const ROMol&)
  Foo(ReadWriteMol)

Fixes this behavior (I think)